+++ davarocr/davar_videotext/tools/track_test.py
@@ -230,8 +230,8 @@ if __name__ == '__main__':
             # calc feature similarty, iou, adjacent matrix to match history track, In YORO, we only use feat_sim_matrix
             # and adja_maxtrix to match, You can adjust to your only task
             feat_sim_matrix = cosine_similarity(track_feature, his_feat_array)
-            iou_matrix = np.zeros([len(cur_frame_pos), len(his_loc_list)], np.float)
-            adja_matrix = np.zeros([len(cur_frame_pos), len(his_loc_list)], np.int)
+            iou_matrix = np.zeros([len(cur_frame_pos), len(his_loc_list)], float)
+            adja_matrix = np.zeros([len(cur_frame_pos), len(his_loc_list)], int)
 
             for cur_idx, cur_bbox in enumerate(cur_frame_pos):
                 cur_poly = test_utils.polygon_from_points(cur_bbox)
@@ -268,7 +268,7 @@ if __name__ == '__main__':
             match_matrix = feat_sim_matrix + (0.1 * adja_matrix + 1e-3)
 
             # aggregate feature similarity matrix and adjacent matrix
-            square_cost_matrix = np.zeros([max_num, max_num], np.float) + 1e-7
+            square_cost_matrix = np.zeros([max_num, max_num], float) + 1e-7
 
             for i in range(feat_sim_matrix.shape[0]):
                 for j in range(feat_sim_matrix.shape[1]):
@@ -285,7 +285,7 @@ if __name__ == '__main__':
 
             # assertion of matching valid , save res to corresponding data and updating history
             ori_his_len = len(his_duration_list)
-            his_matched_matrix = np.zeros([ori_his_len], np.int)
+            his_matched_matrix = np.zeros([ori_his_len], int)
 
             for row_ind_idx, row_item in enumerate(row_ind):
                 cur_idx = row_item
